implementacion de formulario por html

+++ db/migrate/20200324173027_create_histories.rb
@@ -1,8 +1,8 @@
 class CreateHistories < ActiveRecord::Migration[5.2]
   def change
     create_table :histories do |t|
-      t.integer :weight
-      t.integer :height
+      t.float :weight
+      t.float :height
       t.string :observation
       t.references :pet, foreign_key: true
